Unify integration tests fixtures
