+++ scripts/scale-test/tests/integration/test_scale_test.py
@@ -1,29 +1,9 @@
-import json
-import subprocess
 import sys
-import tempfile
 from unittest.mock import patch
 
 import pytest
 
 import scale_test
-from benchmarks.clients import juju
-from benchmarks.models import ClusterInfo, Unit
-
-TEST_JUJU_STATUS_OUTPUT = b"""{"model":{"name":"microk8s","type":"iaas","controller":"mk8s-testing-controller","cloud":"mk8s-testing","region":"Boston","version":"2.9.29","model-status":{"current":"available","since":"26 May 2022 12:58:26+02:00"},"sla":"unsupported"},"machines":{"0":{"juju-status":{"current":"started","since":"26 May 2022 13:02:29+02:00","version":"2.9.29"},"hostname":"juju-c6c89a-0","dns-name":"10.246.154.108","ip-addresses":["10.246.154.108"],"instance-id":"juju-c6c89a-0","machine-status":{"current":"allocating","message":"powering on","since":"26 May 2022 12:58:46+02:00"},"modification-status":{"current":"idle","since":"26 May 2022 12:58:35+02:00"},"series":"focal","network-interfaces":{"ens192":{"ip-addresses":["10.246.154.108"],"mac-address":"00:50:56:09:5c:07","gateway":"10.246.154.1","is-up":true}},"constraints":"arch=amd64 cores=2 mem=4096M root-disk=40960M","hardware":"arch=amd64 cores=2 mem=4096M root-disk=40960M root-disk-source=vsanDatastore"},"1":{"juju-status":{"current":"started","since":"26 May 2022 13:02:29+02:00","version":"2.9.29"},"hostname":"juju-c6c89a-1","dns-name":"10.246.154.111","ip-addresses":["10.246.154.111"],"instance-id":"juju-c6c89a-1","machine-status":{"current":"allocating","message":"powering on","since":"26 May 2022 12:58:46+02:00"},"modification-status":{"current":"idle","since":"26 May 2022 12:58:35+02:00"},"series":"focal","network-interfaces":{"ens192":{"ip-addresses":["10.246.154.111"],"mac-address":"00:50:56:09:5c:07","gateway":"10.246.154.1","is-up":true}},"constraints":"arch=amd64 cores=2 mem=4096M root-disk=40960M","hardware":"arch=amd64 cores=2 mem=4096M root-disk=40960M root-disk-source=vsanDatastore"}},"applications":{"microk8s-node":{"charm":"ubuntu","series":"focal","os":"ubuntu","charm-origin":"charmhub","charm-name":"ubuntu","charm-rev":19,"charm-channel":"stable","exposed":false,"application-status":{"current":"active","since":"26 May 2022 13:02:30+02:00"},"units":{"microk8s-node/0":{"workload-status":{"current":"active","since":"26 May 2022 13:02:30+02:00"},"juju-status":{"current":"idle","since":"26 May 2022 13:02:32+02:00","version":"2.9.29"},"leader":true,"machine":"0","public-address":"10.246.154.108"},"microk8s-node/1":{"workload-status":{"current":"active","since":"26 May 2022 13:02:30+02:00"},"juju-status":{"current":"idle","since":"26 May 2022 13:02:32+02:00","version":"2.9.29"},"leader":false,"machine":"1","public-address":"10.246.154.111"}},"version":"20.04"}},"storage":{},"controller":{"timestamp":"13:03:13+02:00"}}"""  # noqa
-
-
-@pytest.fixture()
-def juju_status_mock():
-    with patch.object(juju, "status") as _status:
-        _status().stdout = TEST_JUJU_STATUS_OUTPUT
-        yield _status
-
-
-@pytest.fixture()
-def subprocess_run_mock():
-    with patch.object(subprocess, "run") as _run:
-        yield _run
 
 
 @pytest.fixture()
@@ -32,20 +12,7 @@ def workload_time():
         yield
 
 
-@pytest.fixture()
-def cluster_json():
-    f = tempfile.NamedTemporaryFile(delete=False)
-    master = Unit(name="name", ip="ip", instance_id="id")
-    cluster = ClusterInfo(master=master, control_plane=[master], workers=[])
-    f.write(json.dumps(cluster.to_dict()).encode())
-    f.file.flush()
-
-    yield f.name
-
-    f.close()
-
-
-def test_main(juju_status_mock, subprocess_run_mock, workload_time, cluster_json):
+def test_main(subprocess_run_mock, workload_time, cluster_json):
     with patch.object(sys, "argv", ["scale_testing", "-c", cluster_json]):
 
         scale_test.main()
